Fix rendering/collision detection bug, move join.js file into parent folder, change colors because I'm bipolar.
The game wasn't rendering properly because I was using Array.slice(-1) when I needed to grab the last element. I've switched it to grab the element instead.

The waitroom folder was causing some circular dependencies and bunches of issues and wasn't really necessary for a single file, so I moved it back into the `web` folder.

I was using the gun colors before, but then I changed my mind and went back to a custom color scheme. I'm bipolar, and I'll probably change it again in the future to be cooler and more gun-like (;
