+++ server/game/expire.js
@@ -3,6 +3,7 @@
 
 var Gun = require('gun/gun');
 var game = require('./index');
+
 function expired(player) {
 	var max, vel = 0.1;
 	max = Math.max.apply(Math, Object.keys(player));
@@ -13,6 +14,7 @@ function expired(player) {
 	instead of a fixed timeout,
 	why not detect wall overlap + latency?
 	That would speed up kicking.
+	Move to collision.js to isomorph.
 */
 setInterval(function () {
 	Gun.obj.map(game.players, function (player, number) {
@@ -20,7 +22,7 @@ setInterval(function () {
 			return;
 		}
 		if (expired(player)) {
-			console.log('removing player', number);
+			console.log('Kicking player', number);
 			game.db.path(number).put(null);
 		}
 	});
